% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scMetalk_abundance.R
\name{calculate_scMetalk_abundance}
\alias{calculate_scMetalk_abundance}
\title{根据代谢酶基因表达计算代谢物丰度}
\usage{
calculate_scMetalk_abundance(seurat_obj, enzyme_file, method = "mean")
}
\arguments{
\item{seurat_obj}{Seurat对象（需包含cell_type列）}

\item{enzyme_file}{代谢酶文件路径（包含Metabolite和Gene列）}

\item{method}{基因整合方法（"mean"或"sum"）}
}
\value{
代谢物丰度矩阵（行：代谢物，列：细胞类型）
}
\description{
根据代谢酶基因表达计算代谢物丰度
}
