% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{load_metabolic_data}
\alias{load_metabolic_data}
\title{加载物种特异性代谢数据}
\usage{
load_metabolic_data(
  data_type = c("transporter", "enzyme", "receptor"),
  species = c("human", "mouse")
)
}
\arguments{
\item{data_type}{数据类型：transporter/enzyme/receptor}

\item{species}{物种：human/mouse}
}
\value{
请求的数据框
}
\description{
加载物种特异性代谢数据
}
