% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scMetalk_sankey.R
\name{plot_scMetalk_alluvial}
\alias{plot_scMetalk_alluvial}
\title{使用桑基图进行细胞间沟通的可视化}
\usage{
plot_scMetalk_alluvial(comm_result, seurat_obj, top_n = 10)
}
\arguments{
\item{seurat_obj}{Seurat对象（用于获取基因表达量）}

\item{top_n}{展示的Top N通讯流}
}
\description{
#' @param comm_result 通讯强度矩阵（来自calculate_metabolic_communication函数）
}
