% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scMetalk_circos.R
\name{plot_scMetalk_circos}
\alias{plot_scMetalk_circos}
\title{使用cirocs图可视化细胞间的代谢通讯}
\usage{
plot_scMetalk_circos(
  comm_data,
  top_n = 20,
  cell_colors = c("#A6CEE3", "#CAB2D6", "#B2DF8A", "#FB9A99", "#FDBF6F", "#1F78B4",
    "#6A3D9A", "#33A02C", "#E31A1C", "#FF7F00"),
  metabolite_palette = "Set1",
  type_colors = c(transporter = "#E31A1C", receptor = "#3498db"),
  track_heights = c(3, 10),
  arrow_params = list(type = "curved", length = 0.15, width_factor = 1, col = "black"),
  legend_position = c(-0.2, 0.95),
  legend_cex = 0.7
)
}
\arguments{
\item{comm_data}{数据框}
}
\value{
可视化circos图
}
\description{
使用cirocs图可视化细胞间的代谢通讯
}
